refactor: move UploadPicker from ui to components
Unlike UI functions that uses Vue internaly, this component is exported
as a Vue component and must be used in a specific Vue version.
